requirements.txt
appnope
backcall
beautifulsoup4
blis
bokeh
bs4
catalogue
certifi
chardet
click 
cloudpickle 
colorama
configparser
crayons
cssselect
cycler
cymem
cytoolz
DateTime
decorator
deprecation
dill
feedfinder2
feedparser
filelock2
frozendict
fsspec 
HeapDict
idna
ipython
ipython-genutils
jedi
jieba3k
Jinja2 
joblib
kiwisolver
locket
lxml
MarkupSafe
matplotlib
msgpack
murmurhash
newspaper3k
nltk
numpy
olefile
packaging 
pandas
parso
partd
pathy
pexpect
pickleshare
Pillow 
polygon-api-client
preshed
prompt-toolkit6
psutil 
ptyprocess
pydantic
Pygments
pyparsing 
python-dateutil 
pytz 
PyYAML
regex
requests
requests-file
selenium
sgmllib3k
six 
smart-open
socketIO-client-nexus
sortedcontainers 
soupsieve
spacy
spacy-legacy
srsly
sseclient7
tblib 
temporal-cache
thinc
tinysegmenter
tldextract
toolz 
tornado 
tqdm
traitlets
typer
typing-extensions
tzlocal
urllib3
wasabi
wcwidth
webdriver-manager
websocket-client
websockets
zict
zopnterface
